fix(audit): v0.29.50-audit1 — deep audit bug fixes (8C+12H+10M)
CRITICAL (8/8):
- C1: inject_fault SystemTrace 3→5 fields (memory_dump + panic_payload)
- C2: inject_fault codegen calls mimi_inject_fault() instead of const_int(0)
- C3: assert_state codegen calls mimi_assert_state() instead of no-op
- C4: LSP protocol byte read errors logged instead of silently discarded
- C5: HTTP fetch timeout failure propagated instead of silently ignored
- C6: JSON number parse failure logged instead of silent 0 return
- C7: Vec::from_raw_parts SAFETY comment added
- C8: SendFilePtr unsafe impl SAFETY comment confirmed present

HIGH (12/17 deduped):
- H3/H9: session_recv type inference i32→i64 (matches runtime mimi_channel_recv)
- H5: SystemKill into_inner() SAFETY comment for mutex poison recovery
- H6/H10: broadcast preserves original value type instead of coercing to 0
- H11: LSP cache file write errors logged instead of silently ignored
- H12: channel recv disconnect logged instead of silent 0 return
- H13: mailbox now_ms() clock failure logged instead of silent 0
- H14: Z3 solver crash distinguished from timeout, crash logged to stderr
- H15: comptime eval error propagated as CompileError instead of eprintln warning
- H17: session.rs same_type copy replaced with delegation to core::helpers

MEDIUM (10/14):
- M1: Fault→Fault transition preserves original SystemTrace from self
- M2: persistent field type default changed from i32 to unit
- M8: Expr::Record mutate param assignment no longer false-positive E0417
- M9: codegen build_in_bounds_gep SAFETY comments added
- M11: LSP stdout flush errors logged instead of .ok() silent ignore
- M13: index builder _ => {} wildcards replaced with explicit Item variants
- M14: mimi_value_is_null() FFI function added to distinguish null from 0

Golden IR files updated for new runtime function declarations.

不变量类别: L1 / L2
测试: flow_features (163 pass), flow dual-backend suite (pass), 3123 lib tests pass

+++ src/core/infer/call/simple.rs
@@ -2086,7 +2086,7 @@ impl<'a> Checker<'a> {
                     if let Some(v) = var {
                         self.set_residual(&v, next);
                     }
-                    return payload_ty.unwrap_or_else(|| Type::Name("i32".into(), vec![]));
+                    return payload_ty.unwrap_or_else(|| Type::Name("i64".into(), vec![]));
                 }
                 Err(e) => {
                     self.emit_code(
@@ -2098,12 +2098,14 @@ impl<'a> Checker<'a> {
             }
         }
         // v0.29.34: when the argument is a plain i64 channel handle (not a
-        // SessionChan<S> variable), session_recv returns i32 (runtime mode).
+        // SessionChan<S> variable), session_recv returns i64 to match the
+        // runtime mimi_channel_recv() which returns i64. Previously returned
+        // i32 which caused checker/codegen divergence (H3-fix).
         let arg_ty = self.infer_expr(&args[0], scopes);
         if same_type(&arg_ty, &Type::Name("i64".into(), vec![])) {
-            return Type::Name("i32".into(), vec![]);
+            return Type::Name("i64".into(), vec![]);
         }
-        Type::Name("i32".into(), vec![])
+        Type::Name("i64".into(), vec![])
     }
 
     pub(in crate::core) fn check_session_close(
